CRUD appellations

+++ Controllers/AlcoholController.cs
@@ -9,13 +9,13 @@ using GuacAPI.Services;
 [ApiController]
 public class AlcoholController : ControllerBase
 {
-    #region fields
+    #region Fields
 
     private IAlcoholService _alcoholService;
 
     #endregion
 
-    #region constructor
+    #region Constructor
 
     public AlcoholController(IAlcoholService alcoholService)
     {
@@ -30,12 +30,18 @@ public class AlcoholController : ControllerBase
     public ActionResult<AlcoholType> GetAllTypes()
     {
         var types = _alcoholService.GetAllTypes();
+
         return Ok(types);
     }
 
     [HttpGet("{id}")]
     public ActionResult<AlcoholType> GetAlcoholTypeById(int id)
     {
+        if (id <= 0)
+        {
+            return BadRequest("Id must be greater than 0");
+        }
+
         var type = _alcoholService.GetAlcoholTypeById(id);
         return Ok(type);
     }
@@ -43,6 +49,11 @@ public class AlcoholController : ControllerBase
     [HttpGet("{label}")]
     public ActionResult<AlcoholType> GetAlcoholByLabel(string label)
     {
+        if (label == "")
+        {
+            return BadRequest("Label cannot be empty");
+        }
+
         var typeLabel = _alcoholService.GetAlcoholByLabel(label);
         return Ok(typeLabel);
     }
@@ -51,12 +62,17 @@ public class AlcoholController : ControllerBase
     public ActionResult<AlcoholType> AddAlcoholType(AlcoholType type)
     {
         var addAlcohol = _alcoholService.AddAlcoholType(type);
-        return Ok(addAlcohol);
+        return Created($"/api/alcohol/{type.AlcoholTypeId}", addAlcohol);
     }
 
     [HttpPut("{id}")]
     public ActionResult<AlcoholType> UpdateAlcoholType(int id, AlcoholType type)
     {
+        if (id <= 0)
+        {
+            return BadRequest("Id must be greater than 0");
+        }
+
         var updateAlcohol = _alcoholService.UpdateAlcoholType(id, type);
         return Ok(updateAlcohol);
     }
@@ -64,6 +80,11 @@ public class AlcoholController : ControllerBase
     [HttpDelete("{id}")]
     public ActionResult<AlcoholType> DeleteAlcoholType(int id)
     {
+        if (id <= 0)
+        {
+            return BadRequest("Id must be greater than 0");
+        }
+
         var deletAlcohol = _alcoholService.DeleteAlcoholType(id);
         return Ok(deletAlcohol);
     }
