+++ Controllers/FurnisherController.cs
@@ -35,6 +35,11 @@ public class FurnisherController : ControllerBase
     [HttpGet("{id}")]
     public IActionResult GetFurnisherById(int id)
     {
+        if (id <= 0)
+        {
+            return BadRequest("Id must be greater than 0");
+        }
+
         var furnisher = this._furnisherService.GetFurnisherById(id);
         return Ok(furnisher);
     }
@@ -42,6 +47,8 @@ public class FurnisherController : ControllerBase
     [HttpGet("{name}")]
     public IActionResult GetFurnisherByName(string name)
     {
+        if (name == "")
+            return BadRequest("Name must be not empty");
         var furnisher = this._furnisherService.GetFurnisherByName(name);
         return Ok(furnisher);
     }
@@ -56,6 +63,7 @@ public class FurnisherController : ControllerBase
     [HttpPut("{id}")]
     public IActionResult UpdateFurnisher(int id, Furnisher furnisher)
     {
+        if (id <= 0) return BadRequest("Id must be greater than 0");
         var updateFurnisher = this._furnisherService.UpdateFurnisher(id, furnisher);
         return Ok(updateFurnisher);
     }
@@ -63,6 +71,7 @@ public class FurnisherController : ControllerBase
     [HttpDelete("{id}")]
     public IActionResult DeleteFurnisher(int id)
     {
+        if (id <= 0) return BadRequest("Id must be greater than 0");
         var deleteFurnisher = this._furnisherService.DeleteFurnisher(id);
         return Ok(deleteFurnisher);
     }
